+++ module/constants/userdata.js
@@ -1,20 +1,48 @@
 import { log } from "../tools.js";
 
 const HERMETIC_FILTER = {
+  techniqueFilter: "",
   formFilter: "",
   levelFilter: 0,
   levelOperator: 0,
-  techniqueFilter: "",
   expanded: false
 };
 
-async function updateUserCache(actorId, list, key, value) {
+const HERMETIC_TOPIC_FILTER = {
+  techniqueFilter: "",
+  formFilter: "",
+  qualityFilter: 0,
+  qualityOperator: 0,
+  expanded: false
+};
+
+const TOPIC_FILTER = {
+  topicFilter: "",
+  typeFilter: "",
+  levelFilter: 0,
+  levelOperator: 0,
+  qualityFilter: 0,
+  qualityOperator: 0,
+  expanded: false
+};
+
+function updateUserCache(actorId, category, list, key, value) {
   let usercache = JSON.parse(sessionStorage.getItem(`usercache-${game.user.id}`));
 
-  usercache[actorId].filters.hermetic[list][key] = value;
+  if (usercache[actorId].filters[category] == undefined) {
+    usercache[actorId].filters[category] = { [list]: { [key]: value } };
+  } else if (usercache[actorId].filters[category][list] == undefined) {
+    usercache[actorId].filters[category][list] = { [key]: value };
+  } else {
+    usercache[actorId].filters[category][list][key] = value;
+  }
   log(false, `Set field ${key} to ${value}`);
   sessionStorage.setItem(`usercache-${game.user.id}`, JSON.stringify(usercache));
-  return;
 }
 
-export { HERMETIC_FILTER, updateUserCache };
+function clearUserCache() {
+  sessionStorage.removeItem(`usercache-${game.user.id}`);
+  ui.notifications.info("User cache has been reset.");
+}
+
+export { TOPIC_FILTER, HERMETIC_FILTER, HERMETIC_TOPIC_FILTER, updateUserCache, clearUserCache };